+++ packages/dev/faas.mjs
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import { main } from './dist/cli.mjs'
+import { main } from './dist/cli/index.mjs'
 
 main(process.argv).then(code => {
   if (code !== 0) process.exit(code)